about us page has info and basic styling
